Rework uniswapV2 stake plugin to support other networks (Ecosystems)
